fixed some dom error
