--- index.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Tic Tac Toe</title>
-</head>
-
-<body>
-  <div class="container">
-    <h1>Tic Tac Toe</h1>
-    <h2 id="p"></h2>
-    <div class="row">
-      <form id="players-form">
-        <div class="input-info">
-          <div class="form-group ml-auto mr-auto">
-            <label class="col-form-label col-form-label-lg" for="inputLarge">Player 1</label>
-            <input class="form-control form-control-lg" type="text" placeholder="name" id="name1" required />
-          </div>
-          <div class="form-group ml-auto mr-auto">
-            <label class="col-form-label col-form-label-lg" for="inputLarge">Player 2</label>
-            <input class="form-control form-control-lg" type="text" placeholder="name" id="name2" required />
-          </div>
-        </div>
-
-        <div class="row">
-          <button type="submit" id="start-game" class="btn btn-primary ml-auto mr-auto">
-            Start Game
-          </button>
-        </div>
-      </form>
-    </div>
-    <div class="row">
-      <button type="button" id="reset-game" onclick="GameBoard.clearBoard();" class="btn btn-primary ml-auto mr-auto">
-        Reset Game
-      </button>
-    </div>
-    <div class="row game-board">
-      <div class="table-data ml-auto mr-auto mt-2"></div>
-    </div>
-  </div>
-  <script src="assets/script.js"></script>
-</body>
-
-</html>
