+++ src/index.html
@@ -4,8 +4,6 @@
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <link rel="stylesheet" href="assets/style.css" />
-  <link rel="stylesheet" href="/assets/bootstrap.min.css" />
   <title>Tic Tac Toe</title>
 </head>
 
@@ -34,15 +32,13 @@
       </form>
     </div>
     <div class="row">
-      <button type="button" id="reset-game" onclick="GameBoard.clearBoard();" class="btn btn-primary ml-auto mr-auto">
+      <button type="button" id="reset-game" class="btn btn-primary ml-auto mr-auto">
         Reset Game
       </button>
     </div>
     <div class="row game-board">
       <div class="table-data ml-auto mr-auto mt-2"></div>
     </div>
-  </div>
-  <script src="assets/script.js"></script>
 </body>
 
 </html>
