+++ src/player.js
@@ -1,0 +1,6 @@
+const players = ((name) => {
+  const token = 'X';
+  return { name, token };
+})();
+
+export default players;
